[CHANGED] LN-001: Created new routes file and moved code from app.js to there
