--- 00-Node-Introduction/first-app.js
@@ -1,3 +1,0 @@
-const fs = require('fs');
-
-fs.writeFileSync('hello.txt', 'Hello from Node.js');
