--- 00-Node-Introduction/hello.txt
@@ -1,1 +1,0 @@
-Hello from Node.js
